+++ heatmap-card.js
@@ -4423,7 +4423,7 @@ window.customCards.push({
     }
 });
 console.info(
-    "%c HEATMAP-CARD %c 2026.9.3 ",
+    "%c HEATMAP-CARD %c 2026.9.6 ",
     "color: black; background: #F2720C; font-weight: 600;",
     "color: black; background: #00a5c9; font-weight: 600;"
 );